move legacy Checkbox
