+++ src/components/legacy/Checkbox/package.json
