+++ src/screens/ImportWallet/ImportWalletLegals.js
@@ -28,7 +28,7 @@ import ContainerWithHeader from 'components/legacy/Layout/ContainerWithHeader';
 import { BaseText } from 'components/legacy/Typography';
 import PrismicDocumentModal from 'components/Modals/PrismicDocumentModal';
 import Button from 'components/legacy/Button';
-import Checkbox from 'components/Checkbox';
+import Checkbox from 'components/legacy/Checkbox';
 import Modal from 'components/Modal';
 
 // utils
